Move common stream code into its own package.
